requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.6
pinecone-client==2.2.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.30.0
six==1.16.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0